refactoring: moved dispaly and keyboard into visualization package, divide display into several entity displays
