+++ handlers/entity_handlers/events.py
@@ -2,9 +2,9 @@ from telegram import Update
 from telegram.ext import CallbackContext
 
 from text import Text
+from fetcher import Route
 from states import States
-from fetcher.fetcher import Route
-from custom_keyboard import CustomKeyboard
+from visualization.custom_keyboard import CustomKeyboard
 from handlers.entity_handlers.base_handler import BaseHandler
